requirements.txt
fastapi
uvicorn
python-pptx
pandas
openpyxl
jinja2
python-multipart
